+++ src/main/java/com/netcracker/app/domain/tariffs/repositories/TariffRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface TariffRepository<E extends Tariff> extends
-        JpaRepository<E, Integer> {
+        JpaRepository<E, Long> {
 
 }
