+++ src/main/java/com/netcracker/app/domain/tariffs/services/TariffService.java
@@ -7,13 +7,13 @@ import java.util.List;
 public interface TariffService<E extends Tariff> {
 
     void add(E e);
-    void delete(int id) throws Exception;
+    void delete(Long id) throws Exception;
     Iterable<E> getAllByName(String name);
     List<E> getAll();
-    E getById(int id);
-    void updateName(String name, int id);
-    void updatePriceOfMonth(double priceOfMonth, int id);
-    void updateDescription(String description, int id);
-    void updateGbInternet(double gbInternet, int id);
-    boolean existsById(int id);
+    E getById(Long id);
+    void updateName(String name, Long id);
+    void updatePriceOfMonth(double priceOfMonth, Long id);
+    void updateDescription(String description, Long id);
+    void updateGbInternet(double gbInternet, Long id);
+    boolean existsById(Long id);
 }
